Split CLI emit-json mode parsing
