+++ src/cli.rs
@@ -1,6 +1,5 @@
 use std::path::Path;
 use std::process;
-use std::str::FromStr;
 
 mod build_graph_execution;
 mod build_graph_loading;
@@ -9,6 +8,7 @@ mod build_graph_targets;
 mod check_emit_commands;
 mod compile;
 mod diagnostics;
+mod emit_json_mode;
 mod execution_commands;
 mod frontend;
 mod json_emit;
@@ -27,6 +27,7 @@ use build_graph_targets::{
 use check_emit_commands::{cmd_check, cmd_emit};
 use compile::{compile_file_to_binary, compile_file_to_c_source, typed_program_to_c_source};
 use diagnostics::{print_diagnostic, print_errors};
+use emit_json_mode::{emit_json_usage, EmitJsonMode};
 use execution_commands::{cmd_build, cmd_build_graph, cmd_run_file, cmd_test};
 use frontend::{graph_frontend, load_module_graph};
 use json_emit::{
@@ -35,98 +36,6 @@ use json_emit::{
     cmd_emit_json_typed,
 };
 use usage::print_usage;
-
-#[derive(Clone, Copy)]
-enum EmitJsonMode {
-    Ast,
-    Symbols,
-    Typed,
-    Diagnostics,
-    BuildGraph,
-    Hir,
-    Mir,
-    Layout,
-    TargetYaml,
-}
-
-impl EmitJsonMode {
-    const AST: &'static str = "ast";
-    const SYMBOLS: &'static str = "symbols";
-    const TYPED: &'static str = "typed";
-    const DIAGNOSTICS: &'static str = "diagnostics";
-    const BUILD_GRAPH: &'static str = "build-graph";
-    const HIR: &'static str = "hir";
-    const MIR: &'static str = "mir";
-    const LAYOUT: &'static str = "layout";
-    const TARGET_YAML: &'static str = "target-yaml";
-
-    const ORDERED: [Self; 9] = [
-        Self::Ast,
-        Self::Symbols,
-        Self::Typed,
-        Self::Diagnostics,
-        Self::BuildGraph,
-        Self::Hir,
-        Self::Mir,
-        Self::Layout,
-        Self::TargetYaml,
-    ];
-
-    const fn as_str(self) -> &'static str {
-        match self {
-            Self::Ast => Self::AST,
-            Self::Symbols => Self::SYMBOLS,
-            Self::Typed => Self::TYPED,
-            Self::Diagnostics => Self::DIAGNOSTICS,
-            Self::BuildGraph => Self::BUILD_GRAPH,
-            Self::Hir => Self::HIR,
-            Self::Mir => Self::MIR,
-            Self::Layout => Self::LAYOUT,
-            Self::TargetYaml => Self::TARGET_YAML,
-        }
-    }
-
-    fn usage() -> String {
-        Self::ORDERED
-            .iter()
-            .map(|mode| mode.as_str())
-            .collect::<Vec<_>>()
-            .join("|")
-    }
-
-    fn gate_message(self) -> Option<&'static str> {
-        match self {
-            Self::Ast
-            | Self::Symbols
-            | Self::Typed
-            | Self::Diagnostics
-            | Self::BuildGraph
-            | Self::Hir
-            | Self::Mir
-            | Self::Layout
-            | Self::TargetYaml => None,
-        }
-    }
-}
-
-impl FromStr for EmitJsonMode {
-    type Err = ();
-
-    fn from_str(value: &str) -> Result<Self, Self::Err> {
-        Self::ORDERED
-            .iter()
-            .copied()
-            .find(|mode| mode.as_str() == value)
-            .ok_or(())
-    }
-}
-
-fn emit_json_usage() -> String {
-    format!(
-        "Usage: zen emit-json <{}> <file.zen>",
-        EmitJsonMode::usage()
-    )
-}
 
 pub fn main() {
     let args: Vec<String> = std::env::args().collect();
